drivers: modem: gsm: Select an MNO profile for U-blox SARA modems
U-blox SARA R4 modems are not meant to run with the factory default
settings. They need an MNO profile other than the default of 0, and
after changing the profile the modem needs to be restarted.

This selects Europe standard as a default. Afterwards it disables
PSM, which is seemingly default in this profile. More preparations
are needed in order to use PSM effectively.

  XXX: This needs some work before it is accepted into Zephyr
  upstream. It does two things that are specific to U-blox SARA: net
  status pin and UMNOPROF. Both of these should be moved to some
  abstraction that lies outside the GSM driver itself.

Signed-off-by: Göran Weinholt <goran.weinholt@endian.se>