drivers: ssd16xx: workaround rounding error in height
The DLC0213CNNG-2 is a 122x250 pixel display. The image data is sent
eight pixels per byte. The image was shifted downwards due to a
rounding error. Unfortunately there are multiple rounding errors and I
don't have time to fix it properly. I set the driver to 128x250 and
added this workaround.

Signed-off-by: Göran Weinholt <goran.weinholt@endian.se>