drivers: display: ssd16xx: Support for partial refresh
This adds support for using the set_contrast() API to decide if the
display does a full or partial refresh. A partial refresh is faster
and happens without display flicker. A full refresh is useful to get
rid of display artifacts.

The controller uses the image from the red RAM as a baseline for
partial refresh, so we keep a copy of the current display contents.
Whether a display supports partial refresh or not depends on the
manufacturer and appropriate values will need to be supplied using
DeviceTree.

Signed-off-by: Göran Weinholt <goran.weinholt@endian.se>