drivers: modem: gsm: Use time since setup when reading RSSI
When reading RSSI during setup we try a few times, because it takes
some time get a usable RSSI value. The code previously used the
uptime, which meant it only worked for the first modem setup. With
recent changes the modem setup is so fast that we don't get a good
RSSI value, so without this fix we're reporting -106 dBm.

Signed-off-by: Göran Weinholt <goran.weinholt@endian.se>